유닛 스탯 포함 버그 수정
1. 식사 효과를 제외한 스탯 조정 수치가 적용되지 않는 현상 수정
2. 숙성 참치회의 효과가 (현재 웨이브 -1)*수치로 적용되던 현상 수정
3. 유닛 스탯 및 데미지 텍스트가 정수형으로 표현되도록 변경
4. 용병 업그레이드 - 장갑 보강의 추가 체력 수치가 %가 아닌 고정값으로 적용되던 현상 수정
5. 소울 보상 공식 변경

+++ Assets/1. Script_New/UI/InGame/DamageText.cs
@@ -22,7 +22,7 @@ public class DamageText : MonoBehaviour
 
     public void SetText(float damage, AttackType attackType)
     {
-        dmg_Text.text = damage.ToString();
+        dmg_Text.text = ((int)damage).ToString();
 
         switch (attackType)
         {
